Make ErrorHandlers module by ActiveSupport::Concerns ActiveSupport::Concerns is awesome! Use it to prevent from creating fat controllers!
